requirements.txt
beautifulsoup4==4.11.1
Flask==2.0.2
Flask_Cors==3.0.10
jinja2==3.0.3
pycryptodome==3.16.0
requests==2.27.1
gevent==22.10.2
